Events ready min
